agent/controller/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
click==8.1.7
colorama==0.4.6
flask==3.1.0
frida==17.2.17
frida-tools==14.4.5
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==5.9.5
Pygments==2.19.2
python-dotenv==1.0.0
wcwidth==0.2.13
websockets==13.1
yarl==1.20.1
